Rearranged the stock controller to minimize what is exposed. All repositories have their own interface now and it is being registered at start up. Had to adjust the schema to account for what is returned from the Yahoo API. Getting quotes at particular times doesn't work yet because it's well after hours, so I have to figure out how to handle the timing for the API.

+++ InStock.Lib/Entities/IQuoteMeta.cs
@@ -4,8 +4,8 @@ namespace InStock.Lib.Entities
     {
         DateTime Date { get; set; }
 
-        decimal Price { get; set; }
+        double Price { get; set; }
 
-        decimal Volume { get; set; }
+        long Volume { get; set; }
     }
 }
