+++ InStock.Lib/Entities/QuoteEntity.cs
@@ -9,9 +9,9 @@ namespace InStock.Lib.Entities
 
         public DateTime Date { get; set; }
 
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
-        public decimal Volume { get; set; }
+        public long Volume { get; set; }
 
         public DateTime CreateOnUtc { get; set; }
     }
