+++ InStock.Lib/Models/QuoteModel.cs
@@ -10,9 +10,9 @@ namespace InStock.Lib.Models
 
         public DateTime Date { get; set; }
 
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
-        public decimal Volume { get; set; }
+        public long Volume { get; set; }
 
         public DateTime CreateOnUtc { get; set; }
     }
